requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
deep-translator>=1.11.4
python-dotenv>=1.0.0
